Avoid override mistake in namespace imports (#16025)
