+++ packages/babel-helpers/src/helpers.ts
@@ -372,53 +372,6 @@ helpers.interopRequireDefault = helper("7.0.0-beta.0")`
   }
 `;
 
-helpers.interopRequireWildcard = helper("7.14.0")`
-  function _getRequireWildcardCache(nodeInterop) {
-    if (typeof WeakMap !== "function") return null;
-
-    var cacheBabelInterop = new WeakMap();
-    var cacheNodeInterop = new WeakMap();
-    return (_getRequireWildcardCache = function (nodeInterop) {
-      return nodeInterop ? cacheNodeInterop : cacheBabelInterop;
-    })(nodeInterop);
-  }
-
-  export default function _interopRequireWildcard(obj, nodeInterop) {
-    if (!nodeInterop && obj && obj.__esModule) {
-      return obj;
-    }
-
-    if (obj === null || (typeof obj !== "object" && typeof obj !== "function")) {
-      return { default: obj }
-    }
-
-    var cache = _getRequireWildcardCache(nodeInterop);
-    if (cache && cache.has(obj)) {
-      return cache.get(obj);
-    }
-
-    var newObj = {};
-    var hasPropertyDescriptor = Object.defineProperty && Object.getOwnPropertyDescriptor;
-    for (var key in obj) {
-      if (key !== "default" && Object.prototype.hasOwnProperty.call(obj, key)) {
-        var desc = hasPropertyDescriptor
-          ? Object.getOwnPropertyDescriptor(obj, key)
-          : null;
-        if (desc && (desc.get || desc.set)) {
-          Object.defineProperty(newObj, key, desc);
-        } else {
-          newObj[key] = obj[key];
-        }
-      }
-    }
-    newObj.default = obj;
-    if (cache) {
-      cache.set(obj, newObj);
-    }
-    return newObj;
-  }
-`;
-
 helpers.newArrowCheck = helper("7.0.0-beta.0")`
   export default function _newArrowCheck(innerThis, boundThis) {
     if (innerThis !== boundThis) {
