+++ packages/babel-plugin-transform-modules-commonjs/test/fixtures/misc/import-ns-frozen-globals/dep.cjs
@@ -1,0 +1,2 @@
+Object.freeze(Object.prototype);
+exports.constructor = 2;
