+++ packages/babel-plugin-transform-modules-commonjs/test/fixtures/misc/import-ns-frozen-globals/exec.mjs
@@ -1,0 +1,2 @@
+import * as ns from "./dep.cjs";
+expect(ns.constructor).toBe(2);
